Convert custom hook party into custom context
